Add files via upload
This is a script  I rewrote to clean up an export file.  The export was supposed to contain 4 fields delimited by a TAB.  The issue was the last field, Comments, contained incorrectly placed CRLFs that needed to be removed.  Luckily, each comment ended with a () or a variation.  But, once again, dealing with a free text field, people enter some comments containing (<junk>).  Testing a new version now to account for that